TASK1_Code/requirements.txt
python=3.11
mamba
numpy
pandas
matplotlib
seaborn
scikit-learn
nltk
transformers
pillow
gensim
streamlit
streamlit_option_menu
pytorch
torchvision
torchaudio
cpuonly